# KAFKA-10579: Make Reflections thread safe to resolve flaky NPE scanning failure