@gharris1727 2023-07-17
I did n=10 runs before and after, scanning ~150 plugin packages (~3.3G total disk space) downloaded from the internet.

The synchronization appears to add on average 0.7% extra execution time (from 44.69s to 44.99s) and increases the standard deviation from 387ms to 1332ms. I think that this is an inconsequential change in execution time, so this change should be safe to merge from a performance standpoint.